backend/requirements.txt
flask
flask-cors
python-dotenv
google-generativeai
google-cloud-vision
pillow
gunicorn
pandas
requests
pytesseract
groq